+++ HISTORY.rst
@@ -3,7 +3,7 @@
 History
 -------
 
-2.1.0 (2019-03-??) - `docs <http://snp-pipeline.readthedocs.io/en/2.0-branch/history.html>`_
+2.1.0 (2019-07-01) - `docs <http://snp-pipeline.readthedocs.io/en/2.1-branch/history.html>`_
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 * Added dependency on jobrunner and qarrayrun Python packages.  These modules were formerly
   part of the CFSAN SNP Pipeline and have been released as separate packages.
@@ -24,6 +24,9 @@ History
   independently or removed from all samples. The default behavior removes dense snp
   regions from all samples, matching the behavior of previous releases.  See :ref:`snp-filtering-label`
   for a detailed explanation.
+* Added the ``--purge`` command line option to save space by removing the intermediate output files
+  (the entire samples directory) upon successful completion of a run.  See the
+  :ref:`cmd-ref-run-snp-pipeline` command reference.